Move AccessType into AssignableExpr
