+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/ASTAssignmentExpressionTest.kt
@@ -1,8 +1,8 @@
 package net.sourceforge.pmd.lang.java.ast
 
 import net.sourceforge.pmd.lang.ast.test.shouldBe
-import net.sourceforge.pmd.lang.java.ast.AccessType.READ
-import net.sourceforge.pmd.lang.java.ast.AccessType.WRITE
+import net.sourceforge.pmd.lang.java.ast.ASTAssignableExpr.AccessType.READ
+import net.sourceforge.pmd.lang.java.ast.ASTAssignableExpr.AccessType.WRITE
 
 /**
  * @author Clément Fournier
