+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/ASTUnaryExpressionTest.kt
@@ -1,8 +1,8 @@
 package net.sourceforge.pmd.lang.java.ast
 
 import net.sourceforge.pmd.lang.java.ast.ASTPrimitiveType.PrimitiveType
-import net.sourceforge.pmd.lang.java.ast.AccessType.READ
-import net.sourceforge.pmd.lang.java.ast.AccessType.WRITE
+import net.sourceforge.pmd.lang.java.ast.ASTAssignableExpr.AccessType.READ
+import net.sourceforge.pmd.lang.java.ast.ASTAssignableExpr.AccessType.WRITE
 import net.sourceforge.pmd.lang.java.ast.BinaryOp.ADD
 import net.sourceforge.pmd.lang.java.ast.IncrementOp.*
 import net.sourceforge.pmd.lang.java.ast.ParserTestCtx.Companion.ExpressionParsingCtx
